+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 # Change Log (vscode-powertools)
 
-## 0.17.0 (February 4th, 2019; ?????)
+## 0.17.0 (February 4th, 2019; script arguments)
 
-* added `extension` property to [ScriptArguments](https://egodigital.github.io/vscode-powertools/api/interfaces/_contracts_.scriptarguments.html) interface
+* added `extension` and `state` properties to [ScriptArguments](https://egodigital.github.io/vscode-powertools/api/interfaces/_contracts_.scriptarguments.html) interface
 * fixes
 
 ## 0.16.0 (February 2nd, 2019; code execution)